add config and profile for the FritzBox7360(SL).
The FritzBox is powerd by
lantiq-xrx200 System on Chip and an new traget for the firmware.

add configs/lantiq-xrx200.config
add profiles/lantiq-xrx200.profiles

fritz-tffs and fritz-caldata are  dependencies for the Fritzbox7360(SL).

kmod-ltq-adsl ... and similar kmod's and packages are dependencies
for the integratet Modem and should working (not tested).